refactor(sputnik): move usize primitives to js base
