+++ src/sputnik/src/hooks/js/impls/mod.rs
@@ -1,7 +1,5 @@
 mod db;
 mod hooks;
 mod list;
-mod primitives;
 mod shared;
 mod storage;
-mod utils;
