+++ src/sputnik/src/hooks/js/sdk/db/count_collection_docs_store.rs
@@ -1,5 +1,5 @@
-use crate::hooks::js::types::primitives::JsUsize;
 use crate::hooks::js::types::shared::JsCollectionKey;
+use crate::js::types::primitives::JsUsize;
 use junobuild_satellite::count_collection_docs_store as count_collection_docs_store_sdk;
 use rquickjs::{Ctx, Error as JsError, Exception, Result as JsResult};
