+++ src/sputnik/src/js/mod.rs
@@ -1,8 +1,10 @@
 mod apis;
 pub mod constants;
 mod dev;
+mod js_utils;
 pub mod module;
+mod primitives;
 pub mod runtime;
-mod utils;
 
 pub mod types;
+pub mod utils;
