+++ src/sputnik/src/js/primitives/types.rs
@@ -1,0 +1,4 @@
+pub mod primitives {
+    #[derive(Clone)]
+    pub struct JsUsize(pub usize);
+}
