+++ src/sputnik/src/js/types.rs
@@ -1,1 +1,2 @@
 pub use crate::js::apis::types::candid;
+pub use crate::js::primitives::types::primitives;
